bazaar: BufferStream: some fixes
git-svn-id: svn://ultimatepp.org/upp/trunk@4998 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ bazaar/BufferStream/BufferStream.h
@@ -18,9 +18,9 @@ protected:
 	Vector<byte> data;
 
 public:
-	void           Open(Vector<byte> & d);
+	void           Open(Vector<byte>& d);
 	void           Create();
-	void           Reserve(int n)            { SetSize((int)GetSize() + n); }
+	void           Reserve(int n)            { SetSize(GetSize() + n); }
 	void           Crop()                    { SetSize(GetPos()); }
 
 	Vector<byte>   GetResult();
